+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.3.0 (2018-11-05)
+
+* Added ``SwappingSerializerMixin``.
+  Allows to declaratively swap fields any field instances in child serializers.
+* Added ``DoubleAsStrJsonEncoder`` for handicapped API clients.
+
 0.2.3 (2017-10-18)
 ~~~~~~~~~~~~~~~~~~